app-chromium: file_web_data::_load_autofill_entries using outer join app-chromium: profile.add_web_data_file implemented app-chromium: file_web_data database schema documentation app-chromium: file_web_data::_load_credit_cards (stub)

+++ src/libmobius_core/bytearray.cpp
@@ -37,11 +37,12 @@ std::string
 to_hex (std::uint64_t value, unsigned int siz)
 {
     char buffer[32] = {0};
+
     snprintf (
         buffer,
         sizeof (buffer),
         "%0*llx",
-        siz,
+        static_cast<int>(siz),
         static_cast<unsigned long long> (value)
     );
